pyproject.toml
[tool.poetry]
name = "homeworks_drf"
version = "0.1.0"
description = ""
authors = ["AlexanderPshenichnyy <ter.anti@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.2"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
d = "^0.2.2"
coverage = "^7.4.3"
django-cors-headers = "^4.3.1"

[tool.poetry.dev-dependencies]
drf-yasg = "^1.21.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
